pyproject.toml
[project]
name = "spoo-qr"
version = "1.0.0"
description = "Open-Source QR Code Generator API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.135.1",
    "jinja2>=3.1.6",
    "pillow>=12.1.1",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "qrcode[pil]>=8.2",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.42.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
addopts = [
    "--import-mode=importlib",
    "-n", "auto",
    "--cov=.",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html:htmlcov",
    "--cov-config=pyproject.toml",
    "-p", "no:randomly",
]

[tool.coverage.run]
source = ["."]
parallel = true
omit = [
    "tests/*",
    "*/test_*",
    "*/conftest.py",
    ".venv/*",
    "main.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
    "pass",
    "\\.\\.\\.",
]
fail_under = 0
show_missing = true
skip_covered = true

[dependency-groups]
dev = [
    "anyio[trio]>=4.0.0",
    "httpx>=0.28.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.0",
    "pytest-clarity>=1.0.1",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.14.1",
    "pytest-randomly>=4.0.1",
    "pytest-sugar>=1.1.1",
    "pytest-xdist>=3.8.0",
    "ruff>=0.11.11",
]
